% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARMA-X.R
\name{CARMAX}
\alias{CARMAX}
\title{CARMA-X}
\usage{
CARMAX(
  z.list,
  ld.list,
  lambda.list = NULL,
  output.labels = ".",
  label.list = NULL,
  LD.estimation = T,
  effect.size.prior = "Spike-slab",
  rho.index = 0.99,
  BF.index = 10,
  outlier.hypo.form = "New.way",
  Max.Model.Dim = 2e+05,
  all.iter = 3,
  all.inner.iter = 10,
  epsilon.threshold = 1e-04,
  multi.ethnic = "admixed",
  num.causal = 10,
  y.var = 1,
  tau = 0.04,
  outlier.switch = T,
  outlier.BF.index = 1/3.2,
  purity.index = 0.5
)
}
\arguments{
\item{z.list}{Input list of the summary statistics of the testing loci, and each element of the list is the summary statistics of each ancestry of the admixed cohort.}

\item{ld.list}{Input list of the LD correlation matrix of the testing loci, and each element of the list is the LD matrix of each ancestry of the admixed cohort. The last element of the list is the overall LD matrix that captures correlations across ancestries.}

\item{lambda.list}{Input list of the hyper-parameter \eqn{\eta} of the testing loci, and each element of the list is the hyper-parameter of each ancestry of the admixed cohort.}

\item{output.labels}{Output directory where output will be written while CARMA is running. Default is the OS root directory ".".}

\item{label.list}{Input list of the names of the testing loci. Default is NULL.}

\item{LD.estimation}{he indicator variable of whether turn on the LD estimation of cross-ancestry correlations. We suggest that the LD estimation is used if using external LD matrix.}

\item{effect.size.prior}{The prior of the effect size, where the 'Spike-slab' prior is the default prior.}

\item{rho.index}{A number between 0 and 1 specifying \eqn{\rho} of the estimated credible sets.}

\item{BF.index}{A number smaller than 1 to specifying the threshold of the Bayes factor of the estimated credible models. The default setting is 3.2.}

\item{Max.Model.Dim}{Maximum number of the top candidate models based on the unnormalized posterior probability.}

\item{all.iter}{Maximum iterations for EM algorithm to run.}

\item{all.inner.iter}{Maximum iterations for Shotgun algorithm to run per iteration within EM algorithm.}

\item{epsilon.threshold}{Convergence threshold measured by average of Bayes factors.}

\item{num.causal}{The maximum number of causal variants assumed per locus, which is 10 causal SNPs per locus by default.}

\item{y.var}{The input variance of the summary statistics, the default value is 1 as the summary statistics are standardized.}

\item{outlier.switch}{The indicator variable of whether turn on the outlier detection. We suggest that the detection should always turn on if using external LD matrix.}

\item{outlier.threshold}{The Bayes threshold of the hypothesis testing of determining outliers, which is 10 by default.}
}
\value{
The form of the return is a list, for each list:
\itemize{
\item pip - The posterior inclusion probability.
\item Credibleset - The information regarding the credible set given a threshold \eqn{\rho}.
\item Credible model - The information regarding the credible model given a threshold  of the Bayes factor.
\item Outliers - The information regarding the detected outliers.
}
}
\description{
Implements the CARMA-X model, a Bayesian fine-mapping method specifically designed for admixed populations.
The model identifies putative causal variants at GWAS loci using summary statistics of SNPs within the loci of interest, 
along with corresponding ancestral LD matrices for admixed fine-mapping and an estimated variance of traits. 
CARMA-X includes an LD estimation procedure to correct for inconsistencies introduced by external LD reference panels, 
which is important for admixed populations where reference LD structures may be less accurate to the results of local ancestry inferences on admixed cohorts. 
Additionally, CARMA-X offers a robust outlier detection procedure that resolves discrepancies between the summary statistics and the LD matrix derived from reference panels.
}
\details{
The function performs a Bayesian fine-mapping method.
}
\examples{
# Example 
set.seed(1)
n = 400
p = 500
beta = rep(0,p)
beta[1] = 1
X = matrix(rnorm(n*p),nrow = n,ncol = p)
X = scale(X,center = TRUE,scale = TRUE)
y = drop(X \%*\% beta + rnorm(n))
SS=compute_summary_statistics(X,y)
z.list<-list()
z.list[[1]]<-(SS$betahat/SS$sebetahat)
ld.list<-list()
ld.list[[1]]<-cov(X)
lambda.list<-list()
lambda.list[[1]]<-1/sqrt(p)
CARMA.result<-CARMA_fixed_sigma(z.list,ld.list=ld.list,
lambda.list = lambda.list,effect.size.prior='Hyper-g')
}
