% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{utility}
\alias{utility}
\title{Utility function for processing z and ld matrices}
\usage{
utility(z.list, ld.list)
}
\arguments{
\item{z.list}{A list of z-score summary statistics for different ancestries.}

\item{ld.list}{A list of LD matrices for different ancestries.}
}
\value{
A combined LD matrix for all ancestries.
}
\description{
This function takes summary statistics and LD matrices as input and computes 
a joint LD matrix used for the CARMA-X fine-mapping model.
}
